DDP-5615 | moved family member related class to participant.data package since those classes are hardly related to participant data
