+++ src/main/java/org/broadinstitute/dsm/model/participant/data/AddFamilyMemberPayload.java
@@ -1,4 +1,4 @@
-package org.broadinstitute.dsm.model.familymember;
+package org.broadinstitute.dsm.model.participant.data;
 
 
 import java.util.Optional;
