+++ src/main/java/org/broadinstitute/dsm/route/familymember/AddFamilyMemberRoute.java
@@ -10,8 +10,8 @@ import org.broadinstitute.dsm.db.User;
 import org.broadinstitute.dsm.db.dao.ddp.instance.DDPInstanceDao;
 import org.broadinstitute.dsm.db.dao.participant.data.ParticipantDataDao;
 import org.broadinstitute.dsm.model.participant.data.NewParticipantData;
-import org.broadinstitute.dsm.model.familymember.AddFamilyMemberPayload;
-import org.broadinstitute.dsm.model.familymember.FamilyMemberDetails;
+import org.broadinstitute.dsm.model.participant.data.AddFamilyMemberPayload;
+import org.broadinstitute.dsm.model.participant.data.FamilyMemberDetails;
 import org.broadinstitute.dsm.security.RequestHandler;
 import org.broadinstitute.dsm.statics.DBConstants;
 import org.slf4j.Logger;
