+++ src/test/java/org/broadinstitute/dsm/route/familymember/AddFamilyMemberRouteTest.java
@@ -15,8 +15,8 @@ import org.broadinstitute.dsm.db.dao.participant.data.ParticipantDataDao;
 import org.broadinstitute.dsm.db.dao.user.UserDao;
 import org.broadinstitute.dsm.db.dto.ddp.instance.DDPInstanceDto;
 import org.broadinstitute.dsm.db.dto.participant.data.ParticipantDataDto;
-import org.broadinstitute.dsm.model.familymember.AddFamilyMemberPayload;
-import org.broadinstitute.dsm.model.familymember.FamilyMemberDetails;
+import org.broadinstitute.dsm.model.participant.data.AddFamilyMemberPayload;
+import org.broadinstitute.dsm.model.participant.data.FamilyMemberDetails;
 import org.broadinstitute.dsm.model.participant.data.NewParticipantData;
 import org.junit.AfterClass;
 import org.junit.Assert;
